identify_multipel_face

+++ face/src/routers/cache_routes.py
@@ -3,7 +3,7 @@ from fastapi import APIRouter, Request
 from pydantic import BaseModel, Field
 from connection.redis_conn import get_redis_connection, download_image
 from connection.download import DecodingImage
-from connection import create_vector_index, add_face_to_index, identify_face_imread_with_vector_search, dump_index_face, find_face_internal, add_face_to_index_base64
+from connection import create_vector_index, add_face_to_index, identify_face_imread_with_vector_search, dump_index_face, find_face_internal, add_face_to_index_base64, identify_multipel_face_imread_with_vector_search
 from dotenv import load_dotenv
 
 load_dotenv()
@@ -73,6 +73,25 @@ async def find_face_by_base64_single(item: FaceRequest):
       }
   image_path = DecodingImage(base64_str)
   result = identify_face_imread_with_vector_search(image_path)
+  if os.path.exists(image_path):
+     os.remove(image_path)
+  return {
+          "status": True,
+          "message": "Wajah ditemukan",
+          "data": result
+          }
+
+@router.post("/face/findIDs/base64")
+async def find_face_by_base64_multiple(item: FaceRequest):
+  base64_str = item.base64
+  if not base64_str:
+      return {
+          "status": False,
+          "message": "Data base64 tidak ditemukan",
+          "data": None
+      }
+  image_path = DecodingImage(base64_str)
+  result = identify_multipel_face_imread_with_vector_search(image_path)
   if os.path.exists(image_path):
      os.remove(image_path)
   return {
